Huge changes
*Some serious commenting
*Adjust hooking

+ECBattleStatus.Heat
+ECBattleStatus.RPGCommand
+ECBattleStatus.ActionCommand
+ECBattleStatus.ClearCommand()
+ECBattleStatus.SetSuperArmor(bool)

+ECBattleTargetDecide
+ECBattleTargetDecide.SetTarget(FighterID)

+ECMotion.RequestRagdoll(RequestRagdollOptions)

+ECRenderCharacter.BattleTransformationOn()
+ECRenderCharacter.BattleTransformationOff()

!!!!!!!KIWAMI 2 ONLY!!!!!!!
+ECSoldierInfo.HP
+ECSoldierInfo.MaxHP
+ECSoldierInfo.PowerRatio
+ECSoldierInfo.Attach(Character)
+ECSoldierInfo.Initialize(CharacterNPCSoldierPersonalDataID)
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

+Asset.GetArmsCategory(AssetID)

+AssetUnit.AssetID

+AuthPlay.TickToFrame(uint)
+AuthPlay.FrameToTick(float)

+AuthPlay.GetGameFrame()
+AuthPlay.SetGameFrame()
+AuthPlay.GetEndFrame()
+AuthPlay.GetGameTick()
+AuthPlay.SetGameTick(uint)
+AuthPlay.GetEndTick()
+AuthPlay.GetCurrentPageIndex()

+Battle.PlayEffect(ParticleID, Matrix4x4)

!!!!!!!KIWAMI 2 ONLY!!!!!!!
+Character.RequestStartFighter()
+Character.WarpPosAndAngle(Vector4 position, float angle)
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

+Character.SetCommandSet(BattleCommandSetID commandSet)
+CharacterBase.IsRagdoll()

+Fighter.GetWeapon(AttachmentCombinationID)
+Fighter.IsSync()
+Fighter.Equip(ItemID, AttachmentCombinationID)

+HumanModeManager.CommandsetModel
+HumanModeManager.ToAttackMode(FighterCommandID)
+HumanModeManager.ToEquipWeapon(AssetID weapon)
+HumanModeManager.ToEndReady()
+HumanModeManager.IsDown()

+Item.GetAssetID(ItemID item)

+ParticleManager.IsLoadedRaw(ParticleID)

+Player.GetHeatNow(PlayerID)
+Player.GetHeatMax(PlayerID)
+Player.SetHeatNow(PlayerID)

+RPG.GetJobCommandSetID(PlayerID, RPGJobID)

+Scene.ConfigID

+SoundManager.PlayBGM(ushort, ushort)

+UI.Create(uint sceneID, uint targetID)

+UIHandleBase.PlayAnimationSet(uint id)

+Party.GetEquipItemId(PlayerID, PartyEquipSlotID)

+++ DELibrary.NET/Enums/BattleCommandSetID.cs
@@ -2,7 +2,7 @@
 
 namespace DragonEngineLibrary
 {
-    public enum BattleCommandSetID : uint
+    public enum BattleCommandSetID : ushort
     {
         invalid,         // constant 0x0
         Default,         // constant 0x1
